+++ src/db/askEstimateOptions.ts
@@ -1,6 +1,6 @@
 import { pgTable, text } from "drizzle-orm/pg-core";
-import { askEstimate } from "./askEstimate";
 import { options } from "./options";
+import { askEstimate } from "./ask_estimate";
 
 export const askEstimateOptions = pgTable("ask_estimate_options", {
   askEstimateId: text("askEstimateId").references(() => askEstimate.id, {
