+++ src/db/index.ts
@@ -5,7 +5,6 @@ export * from "./verificationToken";
 export * from "./verificators";
 export * from "./allergies";
 export * from "./alerts";
-export * from "./askEstimate";
 export * from "./company";
 export * from "./employeeCompany";
 export * from "./intolerences";
@@ -27,3 +26,4 @@ export * from "./estimateOptions";
 export * from "./invoiceOptions";
 export * from "./options";
 export * from "./refundStage";
+export * from "./ask_estimate";
